requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-together>=0.1.0
langgraph>=0.0.40
langsmith>=0.1.0
sqlalchemy>=2.0.0
pandas>=2.0.0
streamlit>=1.28.0
python-dotenv>=1.0.0
tabulate>=0.9.0
requests>=2.31.0
together>=0.2.7 